Rename TypeTraitExpr -> TypeTraitExpression and fix some small issues
