+++ OOModel/src/allOOModelNodes.h
@@ -74,7 +74,7 @@
 #include "expressions/ConditionalExpression.h"
 #include "expressions/ThrowExpression.h"
 #include "expressions/LambdaExpression.h"
-#include "expressions/TypeTraitExpr.h"
+#include "expressions/TypeTraitExpression.h"
 
 #include "expressions/types/TypeExpression.h"
 #include "expressions/types/PrimitiveTypeExpression.h"
